# perf: copy entire file before unpacking