--- a/crates/rattler_package_streaming/Cargo.toml
+++ b/crates/rattler_package_streaming/Cargo.toml
@@ -48,3 +48,10 @@ walkdir = { workspace = true }
 rstest = { workspace = true }
 rstest_reuse = { workspace = true }
 insta = { workspace = true, features = ["yaml"] }
+criterion = { workspace = true, features = ["html_reports", "async_tokio"] }
+uuid = { version = "*", features = ["v4"] }
+
+[[bench]]
+name = "extract"
+harness = false
+