@baszalmstra 2024-12-23
This basically means we are no longer downloading and extracting in parallel. I wonder if that makes a difference with a slower connection.