@baszalmstra 2024-12-23
Do you think it would be possible to implement the server in rust instead? That would make the whole thing more standalone.