@baszalmstra 2024-12-23
Look for StaticFileServer (from top of mind) we already have something similar.