@wolfv 2024-12-23
I think my idea would be that we ideally keep parallel downloading and extraction, but if downloading is much faster than extraction, that we store the remaining data in a (spooled) temporary file, so that download can progress at maximum speed.