@wolfv 2024-12-23
I think what could be also interesting is to try out `async_compression` (uv uses this). `uv` also controls the BufReader size more, and uses a 128Kb buffer (124 * 1024). I think we end up using a default BufReader which has a 8Kb buffer which might lead to less optimal download speeds.