@nichmor 2025-01-02
related to https://github.com/conda/rattler/issues/1007